Trying out writing the temp file for the separate process
This will have to be modified - only tests which we want to run in a separate process should have temp file written, the regular tests can be eval'd. Also, cleanup process needs to be written for the *Temp.php files. For some reason this fails on the PHPUnit level with 'Error: Class 'PHPUnit\TextUI\XmlConfiguration\Loader' not found in Standard input code on line 97'. Need to investigate what happens here.