Update Init.php to change how Laravel is detected
The init currently checks for Laravel by looking for laravel/laravel however a new Laravel 11 install does not include that.

As a result `isLaravelInstalled()` incorrectly returns false

That means it does not add the Laravel stub and so defaults to using phpunit for the base case in TestCase.php. Which then causes an undefined error when running any Feature test, including the example one.